ubahndle-core: extract TrainLabel
