+++ berlin/index.html
@@ -23,7 +23,7 @@
     <meta property="og:image" content="screenshot.png" />
 
     <title>Ubahndle - A daily Berlin rapid transit puzzle game</title>
-    <script type="module" src="src/index.jsx" defer></script>
+    <script type="module" src="src/index.tsx" defer></script>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
