+++ frankfurt/package.json
@@ -5,14 +5,14 @@
   "private": true,
   "type": "module",
   "dependencies": {
+    "@ubahndle/core": "workspace:*",
     "i18next": "^21.8.2",
     "i18next-browser-languagedetector": "^6.1.4",
     "maplibre-gl": "^4.3.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-i18next": "^11.16.9",
-    "semantic-ui-react": "^2.1.1",
-    "@ubahndle/core": "workspace:*"
+    "semantic-ui-react": "^2.1.1"
   },
   "scripts": {
     "dev": "vite",
@@ -37,6 +37,8 @@
     ]
   },
   "devDependencies": {
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
     "@vitejs/plugin-react-swc": "^3.7.0",
     "sass": "^1.77.4",
     "vite": "^5.2.12"
